application is migrated from postgree to mariadb

+++ Jokes/src/main/java/com/jokes/vicevi/entities/Jokes.java
@@ -15,7 +15,7 @@ public class Jokes {
 @Id
 @GeneratedValue(strategy=GenerationType.IDENTITY)
 @Column(name="id")
-private Integer id;
+private Long id;
 
 
 @NotBlank
@@ -30,7 +30,7 @@ private int dislikes;
 
 public Jokes() {}
 
-public Jokes(Integer id,String content, Category category, int likes, int dislikes) {
+public Jokes(Long id,String content, Category category, int likes, int dislikes) {
 	
     this.id=id;
 	this.content = content;
@@ -39,11 +39,11 @@ public Jokes(Integer id,String content, Category category, int likes, int dislik
 	this.dislikes = dislikes;
 }
 
-public Integer getId() {
+public Long getId() {
 	return id;
 }
 
-public void setId(Integer id) {
+public void setId(Long id) {
 	this.id = id;
 }
